[FEAt]: use numpye instead of

+++ misc/objdet_tools.py
@@ -15,7 +15,7 @@
 
 ## general package imports
 import cv2
-import numpy as np
+import numpy as np  
 import math
 from shapely.geometry import Polygon
 
@@ -232,7 +232,7 @@ def project_detections_into_bev(bev_map, detections, configs, color=[]):
         cv2.polylines(bev_map, [corners_int], True, color, 2)
 
         # draw colored line to identify object front
-        corners_int = bev_corners.reshape(-1, 2)
+        corners_int = np.int_(bev_corners.reshape(-1, 2)) # Transform to numpy array in order to manager in a better way
         cv2.line(bev_map, (corners_int[0, 0], corners_int[0, 1]), (corners_int[3, 0], corners_int[3, 1]), (255, 255, 0), 2)
